improve(plaap): create same-page view flow selector (#286)
* feat: include flow selector

* feat: add empty flow to create consistent splash

* fix: only allow one component to render at a time

Thanks to @tulun for this catch!
