+++ src/views/PreLaunchAirdrop/PreLaunchAirdrop.tsx
@@ -1,78 +1,39 @@
-import styled from "@emotion/styled";
 import Footer from "components/Footer";
-import { TitleSection } from "./components";
-import { ReactComponent as IIcon } from "assets/sample-airdrop-icon.svg";
 
 import {
   BackgroundLayer,
   ContentWrapper,
   Wrapper,
 } from "./PreLaunchAirdrop.styles";
-import AirdropCard from "./components/AirdropCard";
 import usePreLaunchAirdrop from "./usePreLaunchAirdrop";
 import TravellerFlow from "./components/TravellerFlow";
+import { SplashFlow } from "./components/SplashFlow";
+import { EligibilityFlow } from "./components/EligibilityFlow";
 const PreLaunchAirdrop = () => {
-  const { showTravellerFlow } = usePreLaunchAirdrop();
+  const { activePageFlow } = usePreLaunchAirdrop();
+
+  let activePageComponent: JSX.Element;
+  switch (activePageFlow) {
+    case "eligibility":
+      activePageComponent = <EligibilityFlow />;
+      break;
+    case "traveller":
+      activePageComponent = <TravellerFlow />;
+      break;
+    case "splash":
+      activePageComponent = <SplashFlow />;
+      break;
+    default:
+      activePageComponent = <></>;
+  }
+
   return (
     <Wrapper>
       <BackgroundLayer />
-
-      {!showTravellerFlow && (
-        <ContentWrapper>
-          <TitleSection />
-          <TestDiv>
-            <AirdropCard
-              Icon={TestIcon}
-              check={"undetermined"}
-              title="Across Community Member"
-              description="Community members can check eligibility for the ACX airdrop by connecting their Discord account to an Ethereum wallet."
-            />
-            <AirdropCard
-              Icon={TestIcon}
-              check={"undetermined"}
-              title="Across Liquidity Provider"
-              description="Community members can check eligibility for the ACX airdrop by connecting their Discord account to an Ethereum wallet."
-              externalLink="/"
-            />
-            <AirdropCard
-              Icon={TestIcon}
-              check={"eligible"}
-              title="Across Community Member"
-              description="Community members can check eligibility for the ACX airdrop by connecting their Discord account to an Ethereum wallet."
-            />
-            <AirdropCard
-              Icon={TestIcon}
-              check={"ineligible"}
-              title="Across Community Member"
-              description="Community members can check eligibility for the ACX airdrop by connecting their Discord account to an Ethereum wallet."
-            />
-            <AirdropCard
-              Icon={TestIcon}
-              title="Across Community Member"
-              description="Community members can check eligibility for the ACX airdrop by connecting their Discord account to an Ethereum wallet."
-            />
-          </TestDiv>
-        </ContentWrapper>
-      )}
-      {showTravellerFlow && (
-        <ContentWrapper>
-          <TravellerFlow />
-        </ContentWrapper>
-      )}
+      <ContentWrapper>{activePageComponent}</ContentWrapper>
       <Footer />
     </Wrapper>
   );
 };
-
-const TestIcon = styled(IIcon)`
-  height: 200px;
-  width: 200px;
-`;
-const TestDiv = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 64px;
-  padding: 64px 0;
-`;
 
 export default PreLaunchAirdrop;
