+++ src/views/PreLaunchAirdrop/components/EligibilityFlow/EligibilityFlow.tsx
@@ -1,0 +1,2 @@
+const EligibilityFlow = () => <></>;
+export default EligibilityFlow;
