+++ src/views/PreLaunchAirdrop/components/EligibilityFlow/index.ts
@@ -1,0 +1,1 @@
+export { default as EligibilityFlow } from "./EligibilityFlow";
