+++ src/views/PreLaunchAirdrop/components/SplashFlow/index.ts
@@ -1,0 +1,1 @@
+export { default as SplashFlow } from "./SplashFlow";
